IM HELPFUL :3

+++ main.py
@@ -82,6 +82,7 @@ class UI:
         self.output_text.insert(tk.END, "Perform an operation to get output.\n")
         self.output_text.config(state=tk.DISABLED)
 
+
         self.clear_data_button = tk.Button(self.data_load_pane, bg=self.COLOR_SCHEME["background-secondary"], fg=self.COLOR_SCHEME["foreground"], text="!!! Delete Comp Data !!!", border=0, borderwidth=0, command=self.reset_comp)
         self.clear_data_button.grid(row=5, column=0, columnspan=2)
     
@@ -203,14 +204,14 @@ class UI:
         graphing_data = {}
 
         for team in self.comparing_teams_list.get(0, tk.END):
-            print(team)
+            #print(team)
             graphing_data[team] = self.graph(comparing, True, team)
         
         LineGraph(graphing_data, "Comparing teams: " + comparing_nice_name)
     
     def add_team_to_comparer(self):
         text = self.compare_team_in.get()
-        print(text)
+        #print(text)
         if text.strip() == "":
             return
         if text in self.comparing_teams_list.get(0, tk.END):
@@ -220,7 +221,7 @@ class UI:
     
     def remove_team_from_comparer(self):
         for i in self.comparing_teams_list.curselection():
-            print(self.comparing_teams_list.get(i))
+            #print(self.comparing_teams_list.get(i))
             self.comparing_teams_list.delete(i)
     
     def FrameWidth(self, event):
@@ -316,12 +317,13 @@ class UI:
         points -= int(match_data["2"]["fouls"])*5
 
         return points
+
+    
     
     def show_team(self, event):
-        
-        
         try:
-            team = event.widget.get(event.widget.curselection()[0]) # gets the selected team in the list
+            #print(int(event.widget.get(event.widget.curselection()[0]).split()[0]))
+            team = event.widget.get(event.widget.curselection()[0]).split()[0] # gets the selected team in the list
             self.mid_pane.event_generate("<Configure>")
             self.mid_pane_canvas.yview_moveto(0)
             self.mid_pane_canvas.grid(row=0, column=1, sticky="nsew", padx=0, pady=0)
@@ -375,18 +377,45 @@ class UI:
         for point in processed_data_points.keys():
             self.loaded_team_data[point][2].config(text=processed_data_points[point])
         self.reload_team_scrollbar(None)
+
+
+
     def reload_team_scrollbar(self, e):
         if e:
             print(e.width)
             #self.mid_pane.configure(width=e.width)
         self.mid_pane_canvas.configure(scrollregion=self.mid_pane_canvas.bbox("all"))
+
     def load_teams_into_selector(self):
         teams = list(self.data.keys())
-        teams.sort()
+        teams = sorted(teams)
+        teamsavgpnts = {}
+        #teams.sort()
         self.teams_pane.delete(0, tk.END)
         for i in range(0, len(teams)):
             team = teams[i]
-            self.teams_pane.insert(tk.END, team)
+            teamsavgpnts[team] = str(self.get_avg_pnts(team))
+
+        sorted_teamsavgpnts = sorted([(float(value), key) for (key, value) in teamsavgpnts.items()], reverse = True)
+        #print(sorted_teamsavgpnts)
+        for t in sorted_teamsavgpnts:
+            self.teams_pane.insert(tk.END, t[1] + " - " + str(t[0]))
+        
+
+    def get_avg_pnts(self, t):
+        # ap: avgpnts, t = team
+        ap = 0
+
+        matches = self.data[t] # The matches of the team
+        for match in list(matches.keys()): # Iterating over the match numbers
+            match_data = matches[match] # Getting the match data
+            if len(list(match_data.keys())) != 0:
+                ap += self.process_points(match_data) # processes the points
+
+        match_count = len(list(matches.keys()))
+        ap /= match_count
+
+        return ap
 
     def get_match(self):
         return self.match_no_selector.get()
@@ -415,6 +444,7 @@ class UI:
             self.load_teams_into_selector()
         except Exception as e:
             self.output_to_device_log(e.with_traceback(None))
+
     def rm_data(self):
         remove_data()
 
@@ -423,11 +453,13 @@ class UI:
             data_file.write(json.dumps(self.data))
     
     def reset_comp(self):
-        with open("data.json", "w+") as data:
-            data.write("{}")
-            self.teams_pane.delete(0, tk.END)
-            self.data = {}
-            self.load_teams_into_selector()
+        sure = askyesno(title='Confirmation', message='Are you sure? This will delete ALL competition data.')
+        if sure:
+            with open("data.json", "w+") as data:
+                data.write("{}")
+                self.teams_pane.delete(0, tk.END)
+                self.data = {}
+                self.load_teams_into_selector()
 
     def run(self):
         self.window.mainloop()
